test(forms): refactor form model tests to three files
